@@ -19,89 +19,88 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.9",
-    "@angular/cdk": "^17.3.9",
-    "@angular/common": "^17.3.9",
-    "@angular/compiler": "^17.3.9",
-    "@angular/core": "^17.3.9",
-    "@angular/forms": "^17.3.9",
-    "@angular/material": "^17.3.9",
-    "@angular/platform-browser": "^17.3.9",
-    "@angular/platform-browser-dynamic": "^17.3.9",
-    "@angular/router": "^17.3.9",
+    "@angular/animations": "^18.0.0",
+    "@angular/cdk": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/material": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
+    "@angular/platform-browser-dynamic": "^18.0.0",
+    "@angular/router": "^18.0.0",
     "@asymmetrik/ngx-leaflet": "^17.0.0",
     "@auth0/angular-jwt": "^5.2.0",
     "@cisstech/nge": "^17.7.1",
-    "@myrmidon/auth-jwt-admin": "^4.1.0",
-    "@myrmidon/auth-jwt-login": "^4.0.1",
-    "@myrmidon/cadmus-api": "^8.0.0",
-    "@myrmidon/cadmus-core": "^8.0.0",
-    "@myrmidon/cadmus-flags-pg": "^8.0.0",
-    "@myrmidon/cadmus-flags-ui": "^8.0.0",
-    "@myrmidon/cadmus-fr-epigraphy-ligatures": "^3.0.0",
-    "@myrmidon/cadmus-graph-pg": "^8.0.0",
-    "@myrmidon/cadmus-graph-ui": "^8.0.0",
-    "@myrmidon/cadmus-item-editor": "^8.0.0",
-    "@myrmidon/cadmus-item-list": "^8.0.0",
-    "@myrmidon/cadmus-item-search": "^8.0.0",
-    "@myrmidon/cadmus-mat-physical-size": "^5.0.0",
-    "@myrmidon/cadmus-part-epigraphy-formula-patterns": "^3.0.0",
-    "@myrmidon/cadmus-part-epigraphy-pg": "^3.0.0",
-    "@myrmidon/cadmus-part-epigraphy-support": "^3.0.0",
-    "@myrmidon/cadmus-part-epigraphy-writing": "^3.0.0",
-    "@myrmidon/cadmus-part-general-pg": "^8.0.0",
-    "@myrmidon/cadmus-part-general-ui": "^8.0.1",
-    "@myrmidon/cadmus-part-geo-asserted-locations": "^4.0.0",
-    "@myrmidon/cadmus-part-geo-asserted-toponyms": "^4.0.0",
-    "@myrmidon/cadmus-part-geo-pg": "^4.0.0",
-    "@myrmidon/cadmus-part-philology-pg": "^8.0.0",
-    "@myrmidon/cadmus-part-philology-ui": "^8.0.0",
-    "@myrmidon/cadmus-preview-pg": "^8.0.0",
-    "@myrmidon/cadmus-preview-ui": "^8.0.0",
-    "@myrmidon/cadmus-profile-core": "^8.0.0",
-    "@myrmidon/cadmus-refs-asserted-chronotope": "^5.0.0",
-    "@myrmidon/cadmus-refs-asserted-ids": "^5.0.1",
-    "@myrmidon/cadmus-refs-assertion": "^5.0.0",
-    "@myrmidon/cadmus-refs-dbpedia-lookup": "^5.0.1",
-    "@myrmidon/cadmus-refs-decorated-counts": "^5.0.0",
-    "@myrmidon/cadmus-refs-decorated-ids": "^5.0.0",
-    "@myrmidon/cadmus-refs-doc-references": "^5.0.1",
-    "@myrmidon/cadmus-refs-external-ids": "^5.0.1",
-    "@myrmidon/cadmus-refs-geonames-lookup": "^5.0.2",
-    "@myrmidon/cadmus-refs-historical-date": "^5.0.0",
-    "@myrmidon/cadmus-refs-lookup": "^5.0.0",
-    "@myrmidon/cadmus-refs-proper-name": "^5.0.1",
-    "@myrmidon/cadmus-refs-viaf-lookup": "^5.0.0",
-    "@myrmidon/cadmus-state": "^8.0.0",
-    "@myrmidon/cadmus-text-block-view": "^5.0.0",
-    "@myrmidon/cadmus-text-ed": "^5.0.1",
-    "@myrmidon/cadmus-text-ed-md": "^5.0.2",
-    "@myrmidon/cadmus-text-ed-txt": "^5.0.0",
-    "@myrmidon/cadmus-thesaurus-editor": "^8.0.0",
-    "@myrmidon/cadmus-thesaurus-list": "^8.0.0",
-    "@myrmidon/cadmus-thesaurus-ui": "^8.0.0",
-    "@myrmidon/cadmus-ui": "^8.0.0",
-    "@myrmidon/cadmus-ui-flags-picker": "^5.0.0",
-    "@myrmidon/cadmus-ui-pg": "^8.0.0",
-    "@myrmidon/ng-mat-tools": "^3.0.0",
-    "@myrmidon/ng-tools": "^3.0.0",
-    "@myrmidon/ngx-dirty-check": "^2.0.0",
-    "@myrmidon/paged-data-browsers": "^1.0.0",
+    "@myrmidon/auth-jwt-admin": "^4.2.0",
+    "@myrmidon/auth-jwt-login": "^4.2.3",
+    "@myrmidon/cadmus-api": "^8.1.0",
+    "@myrmidon/cadmus-core": "^8.1.0",
+    "@myrmidon/cadmus-flags-pg": "^8.1.0",
+    "@myrmidon/cadmus-flags-ui": "^8.1.0",
+    "@myrmidon/cadmus-fr-epigraphy-ligatures": "^3.1.0",
+    "@myrmidon/cadmus-graph-pg": "^8.1.0",
+    "@myrmidon/cadmus-graph-ui": "^8.1.0",
+    "@myrmidon/cadmus-item-editor": "^8.1.0",
+    "@myrmidon/cadmus-item-list": "^8.1.0",
+    "@myrmidon/cadmus-item-search": "^8.1.0",
+    "@myrmidon/cadmus-mat-physical-size": "^5.1.0",
+    "@myrmidon/cadmus-part-epigraphy-formula-patterns": "^3.1.0",
+    "@myrmidon/cadmus-part-epigraphy-pg": "^3.1.0",
+    "@myrmidon/cadmus-part-epigraphy-support": "^3.1.0",
+    "@myrmidon/cadmus-part-epigraphy-writing": "^3.1.0",
+    "@myrmidon/cadmus-part-general-pg": "^8.1.0",
+    "@myrmidon/cadmus-part-general-ui": "^8.1.0",
+    "@myrmidon/cadmus-part-geo-asserted-locations": "^4.1.0",
+    "@myrmidon/cadmus-part-geo-asserted-toponyms": "^4.1.0",
+    "@myrmidon/cadmus-part-geo-pg": "^4.1.0",
+    "@myrmidon/cadmus-part-philology-pg": "^8.1.0",
+    "@myrmidon/cadmus-part-philology-ui": "^8.1.0",
+    "@myrmidon/cadmus-preview-pg": "^8.1.0",
+    "@myrmidon/cadmus-preview-ui": "^8.1.0",
+    "@myrmidon/cadmus-profile-core": "^8.1.0",
+    "@myrmidon/cadmus-refs-asserted-chronotope": "^5.1.1",
+    "@myrmidon/cadmus-refs-asserted-ids": "^5.1.2",
+    "@myrmidon/cadmus-refs-assertion": "^5.1.0",
+    "@myrmidon/cadmus-refs-dbpedia-lookup": "^5.1.1",
+    "@myrmidon/cadmus-refs-decorated-counts": "^5.1.0",
+    "@myrmidon/cadmus-refs-decorated-ids": "^5.1.1",
+    "@myrmidon/cadmus-refs-doc-references": "^5.1.0",
+    "@myrmidon/cadmus-refs-external-ids": "^5.1.1",
+    "@myrmidon/cadmus-refs-geonames-lookup": "^5.1.1",
+    "@myrmidon/cadmus-refs-historical-date": "^5.1.1",
+    "@myrmidon/cadmus-refs-lookup": "^5.1.1",
+    "@myrmidon/cadmus-refs-proper-name": "^5.1.1",
+    "@myrmidon/cadmus-refs-viaf-lookup": "^5.1.0",
+    "@myrmidon/cadmus-state": "^8.1.0",
+    "@myrmidon/cadmus-text-block-view": "^5.1.0",
+    "@myrmidon/cadmus-text-ed": "^5.1.0",
+    "@myrmidon/cadmus-text-ed-md": "^5.1.0",
+    "@myrmidon/cadmus-text-ed-txt": "^5.1.0",
+    "@myrmidon/cadmus-thesaurus-editor": "^8.1.0",
+    "@myrmidon/cadmus-thesaurus-list": "^8.1.0",
+    "@myrmidon/cadmus-thesaurus-ui": "^8.1.0",
+    "@myrmidon/cadmus-ui": "^8.1.0",
+    "@myrmidon/cadmus-ui-flags-picker": "^5.1.0",
+    "@myrmidon/cadmus-ui-pg": "^8.1.0",
+    "@myrmidon/ng-mat-tools": "^3.1.0",
+    "@myrmidon/ng-tools": "^3.1.1",
+    "@myrmidon/paged-data-browsers": "^1.1.0",
     "@types/diff-match-patch": "^1.0.36",
     "diff-match-patch": "^1.0.5",
     "gravatar": "^1.8.2",
     "leaflet": "^1.9.4",
-    "marked": "^9.1.6",
-    "monaco-editor": "^0.48.0",
+    "marked": "^12.0.2",
+    "monaco-editor": "^0.49.0",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.2",
     "zone.js": "~0.14.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular/cli": "~17.3.7",
-    "@angular/compiler-cli": "^17.3.9",
-    "@angular/localize": "^17.3.9",
+    "@angular-devkit/build-angular": "^18.0.1",
+    "@angular/cli": "~18.0.1",
+    "@angular/compiler-cli": "^18.0.0",
+    "@angular/localize": "^18.0.0",
     "@types/jasmine": "~5.1.4",
     "@types/leaflet": "^1.9.12",
     "@types/marked": "^6.0.0",
